Project php table list with poker card!!
Hello! i whant to make a table list order... how ever i really don't know who can i apply or onganize the code.
my english is really terrible, so i have just a few little problem to understud. i thought to mache a underclass or subclass of 
the class Card that is in the script below.
lemmy know if you know something. this is a new exsperience for me. Thanks